- PostRepository에서 addTrack() 메소드가 작동하지 않는 버그 수정
- Category ID타입을 String(varchar)에서 long(bigint)로 변경

+++ blog/src/main/java/kr/co/webcash/service/CategoryService.java
@@ -8,7 +8,7 @@ public interface CategoryService {
 
 	List<Category> listByBlogId(String blogId);
 
-	int findLastIdByBlogId(String blogId);
+	long findLastIdByBlogId(String blogId);
 
 	void save(Category category);
